+++ src/server/scripts/Kalimdor/CavernsOfTime/DragonSoul/boss_morchok.cpp
@@ -357,7 +357,7 @@ class boss_morchok: public CreatureScript
 
                             Talk(ANN_CRYSTAL);
                             Talk(SAY_CRYSTAL);
-                            Position pos = me->GetNearPosition(frand(25.0f, 45.0f), frand(0, 2 * M_PI));
+                            Position pos = me->GetNearPosition(frand(25.0f, 45.0f), frand(0, float(2 * M_PI)));
                             me->CastSpell(pos.GetPositionX(), pos.GetPositionY(), pos.GetPositionZ(), SPELL_RESONATING_CRYSTAL, true);
                             events.ScheduleEvent(EVENT_RESONATING_CRYSTAL, urand(12000, 14000));
                             if (bKohcrom && !bFirstCrystal && pKohcrom)
@@ -541,7 +541,7 @@ class npc_morchok_kohcrom: public CreatureScript
                             break;
                         case EVENT_RESONATING_CRYSTAL:
                         {
-                            Position pos = me->GetNearPosition(frand(25.0f, 45.0f), frand(0, 2 * M_PI));
+                            Position pos = me->GetNearPosition(frand(25.0f, 45.0f), frand(0, float(2 * M_PI)));
                             me->CastSpell(pos.GetPositionX(), pos.GetPositionY(), pos.GetPositionZ(), SPELL_RESONATING_CRYSTAL, true);
                             break;
                         }
