+++ src/server/scripts/Kalimdor/LostCityOfTheTolvir/boss_general_husam.cpp
@@ -204,7 +204,7 @@ struct boss_general_husam : public ScriptedAI
                     _x = x + 3.0f * cos(o);
                     _y = y + 3.0f * sin(o);
                     me->SummonCreature(44711, _x, _y, z, o);
-                    o += M_PI / 2;
+                    o += float(M_PI / 2);
                 }
 
                 DoCast(SPELL_SHOCKWAVE);
