+++ src/server/scripts/Legion/ReturnToKarazhan/boss_shade_of_medivh.cpp
@@ -373,7 +373,7 @@ public:
             {
                 for (uint8 i = 0; i < 5; i++)
                 {
-                    pos = GetCaster()->GetNearPosition(20.0f, frand(-M_PI/3, M_PI/3));
+                    pos = GetCaster()->GetNearPosition(20.0f, frand(float(-M_PI/3), float(M_PI/3)));
                     GetCaster()->SendSpellCreateVisual(GetSpellInfo(), &pos);
                 }
                 tickTimer = 1000;
